requirements/cuda.txt
matplotlib==3.6.2
tensorflow-gpu==2.10.0
